Move components from the  directory
