+++ app/assets/scripts/components/form-elements/number-input.js
