+++ app/assets/scripts/components/list.js
