+++ app/assets/scripts/components/responsive.js
