+++ app/assets/scripts/components/vertical-bar-chart.js
